cleanup and add baseline parallel router
This commit has a bunch of cleanups piggybacked onto the
baseline parallel router:

  * Use RRNodeId where possible
  * Remove dead code and docs left from breadth-first router
  * Fix nondeterminism in parallel tatum
  * Fix performance penalty from NetPinTimingInvalidator

+++ vpr/src/route/annotate_routing.cpp
@@ -62,7 +62,7 @@ vtr::vector<RRNodeId, ParentNetId> annotate_rr_node_nets(const Netlist<>& net_li
                                     describe_rr_node(rr_graph,
                                                      device_ctx.grid,
                                                      device_ctx.rr_indexed_data,
-                                                     (int)size_t(rr_node),
+                                                     rr_node,
                                                      is_flat)
                                         .c_str());
                 } else {
